+++ src/commands/launcherCommand.ts
@@ -28,7 +28,7 @@ export async function addScript(uriList: string[]) {
 }
 
 export async function addArg(script: Script) {
-    const arg = await vscode.window.showInputBox({prompt: 'new argument'});
+    const arg = await vscode.window.showInputBox({ prompt: 'new argument' });
     if (arg) {
         script.args.push(arg);
         scriptService.saveScript();
@@ -43,12 +43,48 @@ export async function deleteArg(arg: ArgItem) {
 }
 
 export async function changeArg(arg: ArgItem) {
-    const newArg = await vscode.window.showInputBox({prompt: 'change argument', value: arg.script.args[arg.argIndex]});
+    const newArg = await vscode.window.showInputBox({ prompt: 'change argument', value: arg.script.args[arg.argIndex] });
     if (newArg) {
         arg.script.args[arg.argIndex] = newArg;
         scriptService.saveScript();
     }
     launcherViewDataProvider.refresh();
+}
+
+export async function launchTerminal(node: NodeItem) {
+    let gresArg = '';
+    if (node.node.gres) {
+        const gresNum = await vscode.window.showQuickPick(Array.from({ length: node.node.gres.totalNum - node.node.gres.usedNum + 1 }, (_, i) => i).map(v => v.toString()), { title: 'Choose number of GRES' });
+        if (!gresNum) { return; }
+        gresArg = "--gres=" + node.node.gres.toIdString() + ":" + gresNum;
+    }
+    let mem = undefined;
+    while (!mem) {
+        mem = await vscode.window.showQuickPick(Array.from({ length: Math.floor((node.node.memory - node.node.allocMemory) / 50) }, (_, i) => (i + 1) * 50).map(v => `${v}G`).concat(['Custom...']), { title: 'Choose memory to alloc' });
+        if (!mem) { return; }
+        if (mem === 'Custom...') {
+            mem = await vscode.window.showInputBox({ prompt: 'Custom memory' });
+        }
+    }
+    let time = undefined;
+    while (!time) {
+        time = await vscode.window.showQuickPick(Array.from({ length: 8 }, (_, i) => i + 1).map(v => `${v}:00:00`).concat(['Custom...']), { title: 'Choose time' });
+        if (!time) { return; }
+        if (time === 'Custom...') {
+            time = await vscode.window.showInputBox({ prompt: 'Custom time' });
+        }
+    }
+    let shell = undefined;
+    while (!shell) {
+        shell = await vscode.window.showQuickPick(['zsh', 'bash', 'Custom...'], { title: 'Choose shell' });
+        if (!shell) { return; }
+        if (shell === 'Custom...') {
+            shell = await vscode.window.showInputBox({ prompt: 'Custom shell' });
+        }
+    }
+    const terminal = vscode.window.createTerminal();
+    terminal.show();
+    terminal.sendText(`srun ${gresArg} --nodelist=${node.node.nodeid} --mem ${mem} -t ${time} --pty ${shell} -i`);
 }
 
 
@@ -61,6 +97,8 @@ export function initLauncherCmd(context: vscode.ExtensionContext) {
     context.subscriptions.push(vscode.commands.registerCommand('slurm--.addArg', addArg));
     context.subscriptions.push(vscode.commands.registerCommand('slurm--.deleteArg', deleteArg));
     context.subscriptions.push(vscode.commands.registerCommand('slurm--.changeArg', changeArg));
+
+    context.subscriptions.push(vscode.commands.registerCommand('slurm--.launchTerminal', launchTerminal));
 
     vscode.commands.executeCommand('slurm--.refreshLauncher');
 }