+++ src/view/resourceView.ts
@@ -1,8 +1,10 @@
 import * as vscode from 'vscode';
 
-import { ResourceGres } from '../models/';
+import { NodeState, ResourceGres } from '../models/';
 import { ListItem, NodeItem } from './components';
 import { resourceService } from '../services';
+
+const emptyGresIcon = new vscode.ThemeIcon('close', new vscode.ThemeColor('gitDecoration.deletedResourceForeground'));
 
 function gresIcon(gres: ResourceGres): vscode.ThemeIcon {
     if (gres.usedNum === 0) { return new vscode.ThemeIcon('circle-filled', new vscode.ThemeColor('gitDecoration.addedResourceForeground')); }
@@ -13,7 +15,9 @@ function gresIcon(gres: ResourceGres): vscode.ThemeIcon {
 function getGroupedNode(): ListItem[] {
     return [...resourceService.groupByGres().values()].map(nodes => {
         if (nodes[0].gres === null) { return new ListItem('No GRES', nodes.map(v => new NodeItem(v)), '${length} nodes', undefined, 'gresList', false); }
-        const rgres = ResourceGres.fromArray(nodes.map(v => v.gres!));
+        const availNode = nodes.filter(v => v.state === NodeState.MIXED || v.state === NodeState.IDLE);
+        if (availNode.length === 0) { return new ListItem(nodes[0].gres.toIdString() + '(0/0)', nodes.map(v => new NodeItem(v)), '${length} nodes', emptyGresIcon, 'gresList', false); }
+        const rgres = ResourceGres.fromArray(availNode.map(v => v.gres!));
         return new ListItem(rgres.toString(), nodes.map(v => new NodeItem(v)), '${length} nodes', gresIcon(rgres), 'gresList', false);
     });
 }